# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "bluebell项目接口文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8084",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/auth/signup": {
            "post": {
                "description": "新用户注册账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamSignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "用户已存在",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/likes": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "对文章或评论点赞",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞"
                ],
                "summary": "点赞",
                "parameters": [
                    {
                        "description": "点赞参数",
                        "name": "like",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LikeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "点赞成功",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "取消对文章或评论的点赞",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞"
                ],
                "summary": "取消点赞",
                "parameters": [
                    {
                        "enum": [
                            "article",
                            "comment"
                        ],
                        "type": "string",
                        "description": "目标类型",
                        "name": "target_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "目标ID",
                        "name": "target_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消成功",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/likes/status": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户对文章或评论的点赞状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞"
                ],
                "summary": "获取点赞状态",
                "parameters": [
                    {
                        "enum": [
                            "article",
                            "comment"
                        ],
                        "type": "string",
                        "description": "目标类型",
                        "name": "target_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "目标ID",
                        "name": "target_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "点赞状态",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseLikeStatus"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/likes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取当前用户的点赞列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞"
                ],
                "summary": "获取用户点赞列表",
                "parameters": [
                    {
                        "enum": [
                            "article",
                            "comment"
                        ],
                        "type": "string",
                        "description": "目标类型",
                        "name": "target_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "点赞列表",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseLikeList"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.LikeRequest": {
            "type": "object",
            "required": [
                "target_id",
                "target_type"
            ],
            "properties": {
                "target_id": {
                    "type": "string",
                    "example": "0"
                },
                "target_type": {
                    "type": "string",
                    "enum": [
                        "article",
                        "comment"
                    ]
                }
            }
        },
        "controller.ResCode": {
            "type": "integer",
            "format": "int64",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010,
                1011,
                1012,
                1013,
                1014
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeInvalidParam",
                "CodeUserExist",
                "CodeUserNotExist",
                "CodeInvalidPassword",
                "CodeServerBusy",
                "CodeNeedLogin",
                "CodeInvalidToken",
                "CodeCategoryNotExist",
                "CodeCategoryExist",
                "CodeCategoryNameInvalid",
                "CodeNoPermissionCategory",
                "CodeArticleAlreadyInCategory",
                "CodeNoPermission",
                "CodeArticleNotExist"
            ]
        },
...
        "controller._ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "msg": {}
            }
        },
...
        "controller._ResponseLikeList": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "list": {},
                        "page": {
                            "type": "integer"
                        },
                        "pages": {
                            "type": "integer"
                        },
                        "size": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "msg": {}
            }
        },
        "controller._ResponseLikeStatus": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "is_liked": {
                            "type": "boolean"
                        }
                    }
                },
                "msg": {}
            }
        },
...
        "controller._ResponseSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {},
                "msg": {}
            }
        },
...
        "models.ParamSignUp": {
            "type": "object",
            "required": [
                "password",
                "re_password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "re_password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}